refactor(e2e): Extract common E2E setup and fixtures to e2e/conftest.py
